requirements.txt
aiogram
playwright
beautifulsoup4
requests
python-dotenv
apscheduler
